+++ pkg/deployment/resources/inspector/apsvc.go
@@ -1,0 +1,126 @@
+//
+// DISCLAIMER
+//
+// Copyright 2025 ArangoDB GmbH, Cologne, Germany
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+// http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+//
+// Copyright holder is ArangoDB GmbH, Cologne, Germany
+//
+
+package inspector
+
+import (
+	"context"
+	"time"
+
+	platformApi "github.com/arangodb/kube-arangodb/pkg/apis/platform/v1beta1"
+	utilConstants "github.com/arangodb/kube-arangodb/pkg/util/constants"
+	"github.com/arangodb/kube-arangodb/pkg/util/errors"
+	"github.com/arangodb/kube-arangodb/pkg/util/k8sutil/inspector/arangoplatformservice"
+	inspectorConstants "github.com/arangodb/kube-arangodb/pkg/util/k8sutil/inspector/constants"
+	"github.com/arangodb/kube-arangodb/pkg/util/k8sutil/inspector/definitions"
+	"github.com/arangodb/kube-arangodb/pkg/util/k8sutil/inspector/generic"
+	"github.com/arangodb/kube-arangodb/pkg/util/k8sutil/inspector/throttle"
+)
+
+func init() {
+	requireRegisterInspectorLoader(arangoPlatformServicesInspectorLoaderObj)
+}
+
+var arangoPlatformServicesInspectorLoaderObj = arangoPlatformServicesInspectorLoader{}
+
+type arangoPlatformServicesInspectorLoader struct {
+}
+
+func (p arangoPlatformServicesInspectorLoader) Component() definitions.Component {
+	return definitions.ArangoPlatformService
+}
+
+func (p arangoPlatformServicesInspectorLoader) Load(ctx context.Context, i *inspectorState) {
+	var q arangoPlatformServicesInspector
+
+	q.v1beta1 = newInspectorVersion[*platformApi.ArangoPlatformServiceList, *platformApi.ArangoPlatformService](ctx,
+		inspectorConstants.ArangoPlatformServiceGRv1Beta1(),
+		inspectorConstants.ArangoPlatformServiceGKv1Beta1(),
+		i.client.Arango().PlatformV1beta1().ArangoPlatformServices(i.namespace),
+		arangoplatformservice.List())
+
+	i.arangoPlatformServices = &q
+	q.state = i
+	q.last = time.Now()
+}
+
+func (p arangoPlatformServicesInspectorLoader) Verify(i *inspectorState) error {
+	return nil
+}
+
+func (p arangoPlatformServicesInspectorLoader) Copy(from, to *inspectorState, override bool) {
+	if to.arangoPlatformServices != nil {
+		if !override {
+			return
+		}
+	}
+
+	to.arangoPlatformServices = from.arangoPlatformServices
+	to.arangoPlatformServices.state = to
+}
+
+func (p arangoPlatformServicesInspectorLoader) Name() string {
+	return "arangoPlatformServices"
+}
+
+type arangoPlatformServicesInspector struct {
+	state *inspectorState
+
+	last time.Time
+
+	v1beta1 *inspectorVersion[*platformApi.ArangoPlatformService]
+}
+
+func (p *arangoPlatformServicesInspector) LastRefresh() time.Time {
+	return p.last
+}
+
+func (p *arangoPlatformServicesInspector) Refresh(ctx context.Context) error {
+	p.Throttle(p.state.throttles).Invalidate()
+	return p.state.refresh(ctx, arangoPlatformServicesInspectorLoaderObj)
+}
+
+func (p *arangoPlatformServicesInspector) Version() utilConstants.Version {
+	return utilConstants.VersionV1Beta1
+}
+
+func (p *arangoPlatformServicesInspector) Throttle(c throttle.Components) throttle.Throttle {
+	return c.ArangoPlatformService()
+}
+
+func (p *arangoPlatformServicesInspector) validate() error {
+	if p == nil {
+		return errors.Errorf("ArangoPlatformServiceInspector is nil")
+	}
+
+	if p.state == nil {
+		return errors.Errorf("Parent is nil")
+	}
+
+	return p.v1beta1.validate()
+}
+
+func (p *arangoPlatformServicesInspector) V1Beta1() (generic.Inspector[*platformApi.ArangoPlatformService], error) {
+	if p.v1beta1.err != nil {
+		return nil, p.v1beta1.err
+	}
+
+	return p.v1beta1, nil
+}